+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-## 0.2.5
+
 Added MIDI timestamp support
+Added MidiDevice info in received packets=
+
+## 0.2.5
+Fix MIDI Session Support on iOS Simulator
 
 ## 0.2.4
 iOS Pod fix